pyproject.toml
[tool.poetry]
name = "pylinketrack"
version = "0.1.0"
description = ""
authors = ["diaszano <luucas.dsantos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.5"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.11.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.19"
mkdocstrings = "^0.22.0"
mkdocstrings-python = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
lint-fix = "blue . && isort ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=linketrack -vv"
post_test = "coverage html"
